refactor(#1802): S1 — extract storage/doctor.rs from the storage/mod.rs monolith (R-05 proof-of-approach)
First step of the #1802 storage-split plan (#1798 R-05). Moves the
~678-LOC doctor/observability-probe region (is_namespace_standard ..
doctor_reflection_totals_by_namespace, incl. sweep_pending_action_timeouts
+ the capability-expansion ledger) VERBATIM into a new PRIVATE
storage::doctor submodule with an itemized `pub use` re-export shim, so
every public path (crate::storage::* / crate::db::*) stays byte-stable
and no caller churns. Zero behavior change (mechanical move; verified
byte-identical against the source region modulo the two `super::`
cross-module call prefixes).

- M-SINGLE-ITEM-PATH: `mod doctor;` stays private — the only public
  paths remain crate::storage::* / crate::db::*.
- M-NO-GLOB-REEXPORTS: itemized shim (visibility drift is review-visible).
- qual_10 lockstep: storage/mod.rs ceiling REDUCED 25_050 -> 24_450
  (actual 24_351) + new 800 row for doctor.rs (actual 698) — "thresholds
  fall, never rise" on a refactor split.
- scripts/ci-test-impact.sh: foundational entry widened to src/storage/*
  (dir-glob, mirrors src/identity/*) so the split never weakens the
  "storage changes force the full suite" invariant; --self-test 8/8 PASS.
- Gates: cargo check --all-targets on default + sal + sal-postgres;
  clippy --all-targets -D warnings -D clippy::pedantic on default + sal;
  fmt; check --examples; qual_10 + qual_6_7 ceiling tests green;
  cargo test --lib storage::tests 293/293 (one UNRELATED pre-existing
  #1465-class isolation flake observed and filed separately).

Base: ff04c9e6af1dc5d8172fc711530e8e2dd654b293
