+++ CHANGELOG.md
@@ -60,6 +60,8 @@ Five-agent adversarial security review of the HTTP/MCP surface (auth & access co
 
 ### Changed
 
+- **Refactor (#1802 R-05, step S1): `storage/mod.rs` doctor/observability probes extracted to `src/storage/doctor.rs`** ([#1802](https://github.com/alphaonedev/ai-memory-mcp/issues/1802), from the [#1798](https://github.com/alphaonedev/ai-memory-mcp/issues/1798) R-05 finding). First step of the storage-monolith decomposition (the Fable-authored plan; proof-of-approach for the remaining steps): the ~678-LOC doctor-probe region (`is_namespace_standard` .. `doctor_reflection_totals_by_namespace`, incl. `sweep_pending_action_timeouts` and the capability-expansion ledger) moved VERBATIM to a new PRIVATE `storage::doctor` submodule with an itemized `pub use` re-export shim, so every public path (`crate::storage::*` / `crate::db::*`) is byte-stable and no caller churns. Zero behavior change. Lockstep: the `qual_10` ceiling for `storage/mod.rs` REDUCED 25_050 → 24_450 (+ a new 800 row for `doctor.rs`), and `scripts/ci-test-impact.sh` widened its foundational entry to `src/storage/*` so the split can never weaken the "storage changes force the full suite" invariant.
+
 - **`ai-memory export` de-silenced: it is a convenience view, not the portability path** ([#1944](https://github.com/alphaonedev/ai-memory-mcp/issues/1944), B_WARN de-silencing, 2×5-agent vote `woaiwndla` / `4d3ea1c5`). The JSON `ai-memory export` command (and its HTTP sibling `GET`-admin export) emits `{memories, links, count, exported_at}` — a **memories + links CONVENIENCE view** that silently omitted the substrate's tamper-evidence + governance spine (governance rules, the append-only revision log, forget tombstones, derivation lineage, per-write attestations, the signed-events audit chain). It now announces that scope instead of dropping it silently:
   - **Stderr WARN (CLI).** `ai-memory export` writes a prominent WARN to **stderr only** (never stdout, so a piped `export > corpus.json` stays valid JSON) naming the omitted signed classes, stating the memories + links convenience scope, and directing to `ai-memory backup` (lossless SQLite `VACUUM INTO`) for integrity-preserving portability and `ai-memory export-forensic-bundle` for the signed crypto spine.
   - **Additive in-band markers (CLI + HTTP).** The export payload gains additive, non-breaking fields — `export_scope="memories+links"`, `portability_complete=false`, `excludes=[governance, revisions, tombstones, lineage, attestations, signed_events]` — so a pipe-to-file consumer that never sees the stderr WARN still learns the scope. The pre-existing `{memories, links, count, exported_at}` shape is **unchanged**; the serialization core (`storage::export_all` / `export_links`) is untouched.
